+++ src/Controls/samples/Controls.Sample/Maui.Controls.Sample-net6.csproj
@@ -6,6 +6,7 @@
     <RootNamespace>Maui.Controls.Sample</RootNamespace>
     <AssemblyName>Maui.Controls.Sample</AssemblyName>
     <IsPackable>false</IsPackable>
+    <RuntimeIdentifier Condition="$(TargetFramework.Contains('-ios')) == true ">iossimulator-x64</RuntimeIdentifier>
   </PropertyGroup>
 
   <PropertyGroup Condition="$(TargetFramework.Contains('-windows')) == true ">
