+++ src/Controls/samples/Controls.Sample/XamlApp.xaml.cs
@@ -26,7 +26,7 @@ namespace Maui.Controls.Sample
 
 		// Must not use MainPage for multi-window
 		protected override Window CreateWindow(IActivationState activationState)
-			=> new Window(Services.GetRequiredService<Page>());
+			=> new Window(new Pages.EditorPage());
 
 		public IServiceProvider Services { get; }
 	}
