+++ src/Core/tests/DeviceTests/Handlers/Editor/EditorHandlerTests.cs
@@ -371,19 +371,5 @@ namespace Microsoft.Maui.DeviceTests
 				EditorHandlerTests.UpdateCursorStartPosition(editorHandler, position);
 			}
 		}
-		
-		[Theory]
-		[InlineData(EditorAutoSizeOption.Disabled)]
-		[InlineData(EditorAutoSizeOption.TextChanges)]
-		public async Task AutoSizeInitializesCorrectly(EditorAutoSizeOption option)
-		{
-			var editor = new EditorStub
-			{
-				AutoSize = option,
-				Text = "Test"
-			};
-
-			await ValidatePropertyInitValue(editor, () => editor.AutoSize, GetNativeAutoSize, editor.AutoSize);
-		}
 	}
 }
